backend/vector_db/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==1.10.13
numpy==1.24.3
faiss-cpu==1.7.4
google-generativeai==0.3.1